+++ src/Pneuma.Core/Database/SqlServer/SqlServerDatabaseDriver.cs
@@ -76,6 +76,7 @@ namespace Pneuma.Core.Database.SqlServer
             Prompts = new PromptMethods(this);
             ChatTurns = new ChatTurnMethods(this);
             ChatFeedback = new ChatFeedbackMethods(this);
+            ChatTurnPerfEvents = new ChatTurnPerfEventMethods(this);
         }
 
         #endregion